Update directshow filter - Separate the registry read and write functions into separate files.
