+++ filter/directshow/include/SideData.hpp
@@ -3,9 +3,8 @@
 
 #include <cstddef>
 
+#include <streams.h>
 #include <guiddef.h>
-#include <unknwn.h>
-#include <winerror.h>
 
 DEFINE_GUID(IID_MediaSideDataHDR, 0x53820dbc, 0xa7b8, 0x49c4, 0xb1, 0x7b, 0xe5, 0x11, 0x59, 0x1a, 0x79, 0xc);
 DEFINE_GUID(IID_MediaSideDataHDRContentLightLevel, 0xed6ae576, 0x7cbe, 0x41a6, 0x9d, 0xc3, 0x7, 0xc3, 0x5d, 0xc1, 0x3e, 0xf9);
