+++ filter/directshow/src/Filter.cpp
@@ -3,23 +3,22 @@
 #include <new>
 #include <string>
 
-#include <wrl/client.h>
 #include <streams.h>
 #include <dvdmedia.h>
 #include <tchar.h>
 #include <initguid.h>
-
-#include "resource.h" // Generated by CMake
+#include <wrl/client.h>
 
 #include "AC/Core.hpp"
 #include "AC/Specs.hpp"
 
+#include "Registry.hpp"
 #include "SideData.hpp"
+
+#include "resource.hpp" // Generated by CMake
 
 #define A2T(s) (sizeof(TCHAR) == sizeof(char) ? (LPCTSTR)(s) : (util::asciiToString<TCHAR>(s).c_str()))
 #define T2A(s) (sizeof(TCHAR) == sizeof(char) ? (LPCSTR)(s) : (util::asciiToString<char>(s).c_str()))
-
-#define gRegArgument (RegArgument::instance())
 
 namespace util
 {
@@ -118,57 +117,6 @@ private:
     static constexpr int StringBufferSize = 512;
 
     bool isInitialized = false;
-};
-
-class RegArgument
-{
-private:
-    RegArgument() noexcept;
-    ~RegArgument() noexcept;
-
-public:
-    double getFactor() noexcept;
-    int getDevice() noexcept;
-    int getLimitWidth() noexcept;
-    int getLimitHeight() noexcept;
-    const TCHAR* getProcessorType() noexcept;
-    const TCHAR* getModelName() noexcept;
-
-    void setFactor(double v) const noexcept;
-    void setDevice(int v) const noexcept;
-    void setLimitWidth(int v) const noexcept;
-    void setLimitHeight(int v) const noexcept;
-    void setProcessorType(const TCHAR* v) const noexcept;
-    void setModelName(const TCHAR* v) const noexcept;
-
-public:
-    static RegArgument& instance() noexcept;
-
-public:
-    static constexpr int ProcessorTypeMaxSize = 32;
-    static constexpr int ModelNameMaxSize = 32;
-
-    static constexpr double FactorDefault = 2.0;
-    static constexpr int DeviceDefault = 0;
-    static constexpr int LimitWidthDefault = 1280;
-    static constexpr int LimitHeightDefault = 720;
-private:
-    static constexpr const TCHAR* ProcessorTypeDefault = TEXT("cpu");
-    static constexpr const TCHAR* ModelNameDefault = TEXT("acnet-hdn0");
-    static constexpr const TCHAR* FactorValueName = TEXT("Factor");
-    static constexpr const TCHAR* DeviceValueName = TEXT("Device");
-    static constexpr const TCHAR* LimitWidthValueName = TEXT("LimitWidth");
-    static constexpr const TCHAR* LimitHeightValueName = TEXT("LimitHeight");
-    static constexpr const TCHAR* ProcessorTypeValueName = TEXT("ProcessorType");
-    static constexpr const TCHAR* ModelNameValueName = TEXT("ModelName");
-private:
-    HKEY key{};
-    double factor{};
-    int device{};
-    int limitWidth{};
-    int limitHeight{};
-    TCHAR ProcessorType[ProcessorTypeMaxSize]{};
-    TCHAR modelName[ModelNameMaxSize]{};
 };
 
 DEFINE_GUID(MEDIASUBTYPE_I420, 0x30323449, 0x0000, 0x0010, 0x80, 0x00, 0x00, 0xaa, 0x00, 0x38, 0x9b, 0x71);
@@ -597,110 +545,4 @@ HRESULT PropertyPage::OnApplyChanges()
     gRegArgument.setModelName(buffer);
 
     return S_OK;
-}
-
-RegArgument::RegArgument() noexcept
-{
-    RegCreateKeyEx(HKEY_CURRENT_USER, TEXT("Software\\Anime4KCPP\\DSFilter"), 0, nullptr,
-        REG_OPTION_NON_VOLATILE, KEY_WRITE | KEY_READ, nullptr, &key, nullptr);
-}
-RegArgument::~RegArgument() noexcept
-{
-    RegCloseKey(key);
-}
-double RegArgument::getFactor() noexcept
-{
-    DWORD size = sizeof(factor);
-    auto data = reinterpret_cast<LPBYTE>(&factor);
-    if (ERROR_SUCCESS == RegQueryValueEx(key, FactorValueName, nullptr, nullptr, data, &size))
-        return factor;
-    else return FactorDefault;
-}
-int RegArgument::getDevice() noexcept
-{
-    DWORD size = sizeof(device);
-    auto data = reinterpret_cast<LPBYTE>(&device);
-    if (ERROR_SUCCESS == RegQueryValueEx(key, DeviceValueName, nullptr, nullptr, data, &size))
-        return device;
-    else return DeviceDefault;
-}
-int RegArgument::getLimitWidth() noexcept
-{
-    DWORD size = sizeof(limitWidth);
-    auto data = reinterpret_cast<LPBYTE>(&limitWidth);
-    if (ERROR_SUCCESS == RegQueryValueEx(key, LimitWidthValueName, nullptr, nullptr, data, &size))
-        return limitWidth;
-    else return LimitWidthDefault;
-}
-int RegArgument::getLimitHeight() noexcept
-{
-    DWORD size = sizeof(limitHeight);
-    auto data = reinterpret_cast<LPBYTE>(&limitHeight);
-    if (ERROR_SUCCESS == RegQueryValueEx(key, LimitHeightValueName, nullptr, nullptr, data, &size))
-        return limitHeight;
-    else return LimitHeightDefault;
-}
-const TCHAR* RegArgument::getProcessorType() noexcept
-{
-    DWORD size = ProcessorTypeMaxSize;
-    auto data = reinterpret_cast<LPBYTE>(ProcessorType);
-    if (ERROR_SUCCESS == RegQueryValueEx(key, ProcessorTypeValueName, nullptr, nullptr, data, &size))
-    {
-        ProcessorType[size - 1] = TEXT('\0');
-        return ProcessorType;
-    }
-    else return ProcessorTypeDefault;
-}
-const TCHAR* RegArgument::getModelName() noexcept
-{
-    DWORD size = ModelNameMaxSize;
-    auto data = reinterpret_cast<LPBYTE>(modelName);
-    if (ERROR_SUCCESS == RegQueryValueEx(key, ModelNameValueName, nullptr, nullptr, data, &size))
-    {
-        modelName[size - 1] = TEXT('\0');
-        return modelName;
-    }
-    else return ModelNameDefault;
-}
-void RegArgument::setFactor(const double v) const noexcept
-{
-    DWORD size = sizeof(factor);
-    auto data = reinterpret_cast<const BYTE*>(&v);
-    RegSetValueEx(key, FactorValueName, 0, REG_BINARY, data, size);
-}
-void RegArgument::setDevice(const int v) const noexcept
-{
-    DWORD size = sizeof(device);
-    auto data = reinterpret_cast<const BYTE*>(&v);
-    RegSetValueEx(key, DeviceValueName, 0, REG_DWORD, data, size);
-}
-void RegArgument::setLimitWidth(const int v) const noexcept
-{
-    DWORD size = sizeof(limitWidth);
-    auto data = reinterpret_cast<const BYTE*>(&v);
-    RegSetValueEx(key, LimitWidthValueName, 0, REG_DWORD, data, size);
-}
-void RegArgument::setLimitHeight(const int v) const noexcept
-{
-    DWORD size = sizeof(limitHeight);
-    auto data = reinterpret_cast<const BYTE*>(&v);
-    RegSetValueEx(key, LimitHeightValueName, 0, REG_DWORD, data, size);
-}
-void RegArgument::setProcessorType(const TCHAR* const v) const noexcept
-{
-    DWORD size = ProcessorTypeMaxSize;
-    auto data = reinterpret_cast<const BYTE*>(v);
-    RegSetValueEx(key, ProcessorTypeValueName, 0, REG_SZ, data, size);
-}
-void RegArgument::setModelName(const TCHAR* const v) const noexcept
-{
-    DWORD size = ModelNameMaxSize;
-    auto data = reinterpret_cast<const BYTE*>(v);
-    RegSetValueEx(key, ModelNameValueName, 0, REG_SZ, data, size);
-}
-
-RegArgument& RegArgument::instance() noexcept
-{
-    static RegArgument object{};
-    return object;
 }
