+++ filter/directshow/src/SideData.cpp
@@ -4,7 +4,6 @@
 #include <new>
 #include <unordered_map>
 
-#include <streams.h>
 #include <malloc.h>
 
 #include "SideData.hpp"
